requirements.txt
certifi==2023.7.22
charset-normalizer==3.2.0
et-xmlfile==1.1.0
idna==3.4
il==0.3
numpy==1.25.2
openpyxl==3.1.2
packaging==23.1
pandas==2.0.3
pdf2image==1.16.3
Pillow==10.0.0
PyPDF2==3.0.1
pytesseract==0.3.10
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
tabulate==0.9.0
tzdata==2023.3
urllib3==2.0.4
